Rename `types` to `type1924`.
Because `types` shadows a standard library package.
