+++ trnsysGUI/type1924/createType1924.py
